aws instance replica

+++ disable-selinux.sh
@@ -1,6 +1,21 @@
 #!/bin/sh
 
-sed -i 's/enforcing/disabled/g' /etc/selinux/config
+sudo yum install -y firewalld
 
-sudo systemctl disable firewalld
-sudo systemctl stop firewalld
+sudo sed -i 's/enforcing/disabled/g' /etc/selinux/config
+
+ENABLED=$(sudo systemctl status firewalld | grep "Active: active" | wc -l)
+
+if [ "$ENABLED" -eq "1" ]
+then
+  echo "Disabling selinux"
+  sudo systemctl disable firewalld
+  sudo systemctl stop firewalld
+else
+	echo "Selinux already disabled"
+fi
+
+sudo setenforce 0
+
+
+echo "OK"
